+++ project/src/main/java/es/iespuertodelacruz/xptrade/controllers/v2/PostRESTControllerV2.java
@@ -11,14 +11,22 @@ import es.iespuertodelacruz.xptrade.dto.input.PostInputDTO;
 import es.iespuertodelacruz.xptrade.dto.output.PostOutputDTO;
 import es.iespuertodelacruz.xptrade.mapper.dto.input.IPostInputDTOMapper;
 import es.iespuertodelacruz.xptrade.mapper.dto.output.IPostOutputDTOMapper;
+import es.iespuertodelacruz.xptrade.shared.security.CustomUserDetails;
+import es.iespuertodelacruz.xptrade.shared.security.JwtService;
 import es.iespuertodelacruz.xptrade.shared.utils.CustomApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static es.iespuertodelacruz.xptrade.shared.security.JwtFilter.authHeader;
+import static es.iespuertodelacruz.xptrade.shared.security.JwtFilter.authHeaderTokenPrefix;
 
 @RestController
 @CrossOrigin
@@ -33,7 +41,6 @@ public class PostRESTControllerV2 {
     private IPostService service;
     private IGameService gameService;
     private IUserService userService;
-
     /**
      * Setters of the post service
      * @param service of the post
@@ -229,6 +236,25 @@ public class PostRESTControllerV2 {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> delete(@PathVariable Integer id) {
+
+        Post dbItem = service.findById(id);
+
+        if(dbItem == null){
+            ResponseEntity.status(HttpStatus.NO_CONTENT)
+                    .body(new CustomApiResponse<>(204, "Not found", null));
+        }
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        CustomUserDetails userDetails = (CustomUserDetails) auth.getPrincipal();
+        String username = userDetails.getUsername();
+        User userDb = userService.findByUsername(username);
+
+
+        if (dbItem != null && userDb.getId() != dbItem.getUser().getId()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(new CustomApiResponse<>(401, "Unauthorized", null));
+        }
 
         boolean deleted = service.delete(id);
